consistent amount_cents field types

+++ db/migrate/20260705110029_create_fin_ops_deposits.rb
@@ -3,7 +3,7 @@ class CreateFinOpsDeposits < ActiveRecord::Migration[8.1]
     create_table :fin_ops_deposits do |t|
       t.references :payer, null: false
       t.uuid :public_id, null: false, default: "uuid_generate_v4()"
-      t.integer :amount_cents, null: false
+      t.bigint :amount_cents, null: false
 
       t.timestamps
     end
