+++ db/migrate/20260705114656_create_cx_money_movements.rb
@@ -5,7 +5,7 @@ class CreateCXMoneyMovements < ActiveRecord::Migration[8.1]
       t.references :sender, index: true
       t.integer :kind, limit: 2, null: false
       t.uuid :reference, null: false
-      t.integer :amount_cents, null: false
+      t.bigint :amount_cents, null: false
       t.integer :state, limit: 2, null: false, default: 1
       t.datetime :initiated_at, null: false
       t.datetime :resolved_at
